move name logic to NamedRegistry
